# openapi.yaml
openapi: 3.0.3
info:
  title: Dogfood 2026 REST API
  version: 1.0.0
  description: Same-origin REST API. Protected operations use the dogfood_session httpOnly cookie.
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /events/{slug}/gallery:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    get:
      tags: [Public participation]
      summary: Search and filter public submitted projects
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: track, in: query, schema: { type: string } }
        - { name: seed, in: query, schema: { type: string } }
      responses:
        '200': { description: Stable-seed ordered public project summaries; no private scores }
...
  /events/{slug}/comments:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    get:
      tags: [Public participation]
      summary: List comments for a submission
      parameters: [{ name: submissionId, in: query, required: true, schema: { type: string } }]
      responses:
        '200': { description: Public comments }
    post:
      tags: [Public participation]
      summary: Add a comment (participant)
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [submissionId, body], properties: { submissionId: { type: string }, body: { type: string, maxLength: 1000 } } }
      responses:
        '201': { description: Comment created }
        '429': { description: Rate limited }
  /events/{slug}/assignments:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    get:
      tags: [Judging]
      summary: List the caller's assignments or event assignments (manager)
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: Assignments with score state }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [Judging]
      summary: Batch assign judges manually or by least-loaded heuristic
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [mode, judgeIds, submissionIds], properties: { mode: { type: string, enum: [manual, auto] }, judgeIds: { type: array, items: { type: string } }, submissionIds: { type: array, items: { type: string } }, perSubmission: { type: integer, default: 2 } } }
      responses:
        '201': { description: Assignment rows }
        '403': { $ref: '#/components/responses/Error' }
...
  /events/{slug}/bulk:
    parameters: [{ $ref: '#/components/parameters/Slug' }]
    get:
      tags: [Operations]
      summary: Export JSON snapshot or selected CSV round-trip dataset (owner/admin)
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: format, in: query, schema: { type: string, enum: [json, csv] } }
        - { name: type, in: query, schema: { type: string, enum: [teams, submissions, scores] } }
      responses: { '200': { description: JSON snapshot or CSV attachment for teams, submissions, or scores } }
    post:
      tags: [Operations]
      summary: Import version 1 teams and submissions (owner/admin)
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [version], properties: { version: { type: integer, enum: [1] }, teams: { type: array, items: { type: object } }, submissions: { type: array, items: { type: object } }, scores: { type: array, items: { type: object } } } }
          text/csv:
            schema: { type: string, description: CSV import; type query parameter selects teams, submissions, or score rows }
      responses: { '201': { description: Import counts }, '400': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dogfood_session
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: JSON error object
      content:
        application/json:
          schema: { type: object, properties: { error: { type: string } } }